requirements.txt
django
django-cors-headers
djangorestframework
django-db-connection-pool[mysql]
mysqlclient
python-dateutil
boto3
django-filter
DateTime
requests
sqlparse
pandas
sqlalchemy
openpyxl
xlsxwriter
PyJWT==1.7.1
pysftp
pytz
drf-spectacular
# reportlab
xhtml2pdf
bcrypt
drf-yasg